move test lightclient methods to test-features mod
